+++ caliper/__init__.py
@@ -33,8 +33,8 @@ install_aliases()
 from builtins import *
 
 __title__ = 'imsglobal_caliper'
-__version__ = '1.1.10.6'
-__build__ = 0x01011006
+__version__ = '1.2.0.0'
+__build__ = 0x01020000
 __author__ = 'IMS Global Learning Consortium, Inc.'
 __license__ = 'LGPLv3'